scripts/naver-auto-poster/requirements.txt
undetected-chromedriver>=3.5.5
selenium>=4.15.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
